+++ frontend/src/metabase/core/hooks/use-form-error-message/use-form-error-message.ts
@@ -1,6 +1,6 @@
 import { useLayoutEffect, useState } from "react";
 import { useFormikContext } from "formik";
-import { useFormState } from "./use-form-state";
+import useFormState from "../use-form-state";
 
 const useFormErrorMessage = () => {
   const { values } = useFormikContext();
