+++ frontend/src/metabase/core/hooks/use-form-state/types.ts
@@ -1,0 +1,6 @@
+export type FormStatus = "pending" | "fulfilled" | "rejected";
+
+export interface FormState {
+  status?: FormStatus;
+  message?: string;
+}
