+++ frontend/src/metabase/core/hooks/use-form-state/use-form-state.ts
@@ -1,0 +1,9 @@
+import { useFormikContext } from "formik";
+import { FormState } from "./types";
+
+const useFormState = (): FormState => {
+  const { status } = useFormikContext();
+  return status ?? {};
+};
+
+export default useFormState;
