+++ frontend/src/metabase/core/hooks/use-form/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./use-form";
