+++ frontend/src/metabase/core/hooks/use-form/types.ts
@@ -1,0 +1,10 @@
+import type { FormikErrors } from "formik";
+
+export interface FormError<T> {
+  data?: FormErrorData<T>;
+}
+
+export interface FormErrorData<T> {
+  errors?: FormikErrors<T>;
+  message?: string;
+}
